+++ unstructured/documents/elements.py
@@ -156,33 +156,6 @@ class FormKeyValuePair(TypedDict):
     confidence: float
 
 
-def _kvform_rehydrate_internal_elements(kv_pairs: list[dict]) -> list[FormKeyValuePair]:
-    from unstructured.staging.base import elements_from_dicts
-
-    # safe to overwrite - deepcopy already happened
-    for kv_pair in kv_pairs:
-        if kv_pair["key"]["custom_element"] is not None:
-            (kv_pair["key"]["custom_element"],) = elements_from_dicts(
-                [kv_pair["key"]["custom_element"]]
-            )
-        if kv_pair["value"] is not None and kv_pair["value"]["custom_element"] is not None:
-            (kv_pair["value"]["custom_element"],) = elements_from_dicts(
-                [kv_pair["value"]["custom_element"]]
-            )
-    return kv_pairs
-
-
-def _kvform_pairs_to_dict(kv_pairs: list[FormKeyValuePair]) -> list[dict]:
-    kv_pairs: list[dict] = copy.copy(kv_pairs)  # deepcopy already happened
-    for kv_pair in kv_pairs:
-        if kv_pair["key"]["custom_element"] is not None:
-            kv_pair["key"]["custom_element"] = kv_pair["key"]["custom_element"].to_dict()
-        if kv_pair["value"] is not None and kv_pair["value"]["custom_element"] is not None:
-            kv_pair["value"]["custom_element"] = kv_pair["value"]["custom_element"].to_dict()
-
-    return kv_pairs
-
-
 class ElementMetadata:
     """Fully-dynamic replacement for dataclass-based ElementMetadata."""
 
@@ -369,6 +342,8 @@ class ElementMetadata:
             elif field_name == "orig_elements":
                 self.orig_elements = elements_from_base64_gzipped_json(field_value)
             elif field_name == "key_value_pairs":
+                from unstructured.documents.form_utils import _kvform_rehydrate_internal_elements
+
                 self.key_value_pairs = _kvform_rehydrate_internal_elements(field_value)
             else:
                 setattr(self, field_name, field_value)
@@ -436,6 +411,8 @@ class ElementMetadata:
         if self.orig_elements is not None:
             meta_dict["orig_elements"] = elements_to_base64_gzipped_json(self.orig_elements)
         if self.key_value_pairs is not None:
+            from unstructured.documents.form_utils import _kvform_pairs_to_dict
+
             meta_dict["key_value_pairs"] = _kvform_pairs_to_dict(self.key_value_pairs)
 
         return meta_dict
@@ -1042,7 +1019,9 @@ class PageNumber(Text):
     category = "PageNumber"
 
 
-class FormKeysValues(Element):
+class FormKeysValues(Text):
+    """An element for capturing Key-Value dicts (forms)."""
+
     category = "FormKeysValues"
